refactor(telemetry): address PR 367 review feedback
- Export OneCollectorLogExporter in library.__all__
- Switch test_deviceid and test_exporter to package-level imports
- Drop dead _MIN_HTTP_TIMEOUT clamp; use _HTTP_TIMEOUT directly
